♻️ refactor: register view has now clean code
Important Changes:

> The form from register is now modularize in another class (registerForm)

> For a better organization of the files, in the widgets files is nos available a file named Forms intended to store all the forms used in the app.

> The name of the util string change to strings_app for a better comprehension.

> Now when retrieving the values from the formfiels, a method to clean white spaces is now working.
