+++ lib/controllers/auth_controller.dart
@@ -1,23 +1,45 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:club_steam_app/utils/Strings.dart';
+import 'package:club_steam_app/utils/strings_app.dart';
 
 class AuthController {
   final FirebaseAuth _auth = FirebaseAuth.instance;
+  late UserCredential userCredential;
+
+  Future<String> registerUser(String email, String password) async {
+    try {
+      userCredential =
+          await FirebaseAuth.instance.createUserWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+      return FormFieldMessage.sucessfulRegistration;
+    } on FirebaseAuthException catch (e) {
+      if (e.code == 'email-already-in-use') {
+        return FormFieldMessage
+            .existingAccount; // Ensure this constant is defined
+      } else {
+        return FormFieldMessage
+            .errorConnectionFirebase; // Ensure this constant is defined
+      }
+    } catch (e) {
+      return "An unexpected error occurred: ${e.toString()}";
+    }
+  }
 
   // Verify the account of the user
   Future<String> loginUserWithPassword(String email, String password) async {
     try {
       await _auth.signInWithEmailAndPassword(email: email, password: password);
-      return LoginMessages.successfulLogin;
+      return FormFieldMessage.successfulLogin;
     } on FirebaseAuthException catch (e) {
       if (e.code == 'user-not-found') {
-        return LoginMessages.wrongEmail;
+        return FormFieldMessage.wrongEmail;
       } else if (e.code == 'wrong-password') {
-        return LoginMessages.wrongPaswword;
+        return FormFieldMessage.wrongPaswword;
       } else if (e.code == 'network-request-failed') {
-        return LoginMessages.noNetworkConnection;
+        return FormFieldMessage.noNetworkConnection;
       } else {
-        return LoginMessages.errorConnectionFirebase;
+        return FormFieldMessage.errorConnectionFirebase;
       }
     }
   }
