+++ lib/utils/strings_app.dart
@@ -1,10 +1,14 @@
 import 'package:flutter/material.dart';
 
 // Messages used on Login View
-class LoginMessages {
+class FormFieldMessage {
   static String wrongEmail = "No se encontro un usuario con dicho correo";
   static String wrongPaswword = "Contraseña incorrecta";
   static String successfulLogin = 'Inicio de Sesión exitoso';
   static String noNetworkConnection = "No hay conexión a internet";
   static String errorConnectionFirebase = 'Error al conectarse al servidor';
+  static String existingAccount = 'Ya existe una cuenta asociada';
+  static String noNullFields = 'Por favor, no deje este campo vacio';
+  static String selectOption = 'Por favor, seleccione una opción';
+  static String sucessfulRegistration = "Usuario registrado exitosamente";
 }
